feat: isolates photoswipe in a mixin
- isolates photoswipe code
- adds `setDataPhotoswipe` validator
